requirements.txt
PySide6>=6.5.0
opencv-python-headless>=4.9.0.80
opencolorio>=2.3.2
qimage2ndarray>=1.10.0
numba>=0.59.1
